updated user authentication,Session management jwt and user validation loggin

+++ backend/MyHabari/v1/src/database/userDb.mjs
@@ -1,5 +1,9 @@
 //app dependancies
 import { PrismaClient } from '@prisma/client';
+import { validationResult } from 'express-validator';
+
+
+
 const prisma = new PrismaClient();
 
 process.on("uncaughtException", (err) => {
@@ -11,7 +15,15 @@ process.on("uncaughtException", (err) => {
 
 export const  UsersList=async (req,res)=>{
   try {
-    const users = await prisma.Users.findMany();
+    const users = await prisma.Users.findMany({
+      //abstracting the password
+      select: {
+        id: true,
+        Name: true,
+        Email: true,
+        Admin: true,
+      }}
+    );
     res.json(users);
   } catch (error) {
     res.status(500).json({ error: error.message });
@@ -20,12 +32,24 @@ export const  UsersList=async (req,res)=>{
 
   //fetching a single user
 export const  user=async (req,res)=>{
+  const errors = validationResult(req)    
+  if(!errors.isEmpty()) {
+      return res.status(400).json({errors: errors.array()})
+  }
+
     try {
       const { query: {User_id } } = req;
       const parsedId = parseInt(User_id);
 
       const user = await prisma.Users.findUnique({
         where: { id: parsedId },
+        //abstracting the password
+        select: {
+          id: true,
+          Name: true,
+          Email: true,
+          Admin: true,
+        },
       });
 
       if (!user) {
@@ -46,12 +70,16 @@ export const AddUser = async (NewUser) => {
       try {
         const createdUser = await prisma.Users.create({
           data: NewUser,
+          select: {
+            id: true,
+            Name: true,
+            Email: true,
+            Admin: true,
+          },
         });
-        console.log(createdUser);
         return createdUser;
       
       } catch (error) {
-        console.error(`Error adding user: ${error.message}`);
         throw error;
       }
     };
@@ -59,6 +87,12 @@ export const AddUser = async (NewUser) => {
 
 
 export const patchUser = async (req, res) => {
+
+  const errors = validationResult(req)    
+  if(!errors.isEmpty()) {
+      return res.status(400).json({errors: errors.array()})
+  }
+
     const { query: {User_id } } = req;
     const parsedId = parseInt(User_id);
 
@@ -75,10 +109,16 @@ export const patchUser = async (req, res) => {
         const updatedUser = await prisma.Users.update({
           where: { id: parsedId }, 
           data: updateData,
+          select: {
+            id: true,
+            Name: true,
+            Email: true,
+            Admin: true,
+          },
         });
         res.status(200).json(updatedUser);
       } catch (error) {
-      res.status(500).json({ error: error.message });
+      res.status(500).json({message:"user not found"});
       }
     };
 
@@ -89,17 +129,37 @@ export const patchUser = async (req, res) => {
 
  //deleting a user   
 export const deleteuser  = async (req,res)=> {
+
+  //vaidating the user id querry params
+  const errors = validationResult(req)    
+  if(!errors.isEmpty()) {
+      return res.status(400).json({errors: errors.array()})
+  }
+
+
   try {
     const { query: {User_id } } = req;
     const parsedId = parseInt(User_id);
-
     const user = await prisma.Users.delete({
       where: { id: parsedId },
     });
-
     res.status(200).json({ message: 'User deleted successfully',user });
   } catch (error) {
-    // Handle potential errors, such as the user not being found
-    res.status(500).json({ error: error.message });
+    res.status(500).json({ message:"user not found" });
   }
 };
+
+
+//check password and password
+export const LogUser = async (user) => {
+  try {
+    const LoggedUser = await prisma.Users.findUnique({
+      where: { Email: user.Email }
+    });
+    return LoggedUser;
+  
+  } catch (error) {
+    throw error;
+  }
+};
+
